arm: cpu: fix starting OFF and booting with NCPUHALT+RESET
We can't set power_state on reset, because the change
in power_state done by GPIO input is wrongly overriden
by the setting in reset.

Also, when NCPUHALT is set low, we do not want to run the CPU
unconditionally, because its RESET input might still be asserted (high),
so we need to check the reset pin, before setting the interrupt mask.